requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via openai
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flasgger==0.9.7.1
    # via -r requirements.in
flask==3.1.0
    # via
    #   -r requirements.in
    #   flasgger
    #   flask-cors
flask-cors==5.0.1
    # via -r requirements.in
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
google-auth==2.40.2
    # via google-genai
google-genai==1.18.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
huggingface-hub==0.31.1
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
jiter==0.9.0
    # via openai
joblib==1.5.0
    # via scikit-learn
jsonschema==4.23.0
    # via flasgger
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mistune==3.1.3
    # via flasgger
mpmath==1.3.0
    # via sympy
mysql-connector-python==9.3.0
    # via -r requirements.in
networkx==3.4.2
    # via torch
numpy==2.2.5
    # via
    #   scikit-learn
    #   scipy
    #   transformers
openai==1.78.1
    # via -r requirements.in
packaging==25.0
    # via
    #   flasgger
    #   huggingface-hub
    #   transformers
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.4
    # via
    #   google-genai
    #   openai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   flasgger
    #   huggingface-hub
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-genai
    #   huggingface-hub
    #   transformers
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
six==1.17.0
    # via flasgger
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
torch==2.7.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.51.3
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   anyio
    #   google-genai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   torch
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.3.0
    # via requests
websockets==15.0.1
    # via google-genai
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
